moved result table and page
