+++ web/frontend/src/pages/Footer.tsx
