+++ web/frontend/src/pages/ResultShow.css
