+++ web/frontend/src/pages/ResultShow.tsx
@@ -3,16 +3,16 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 
-import { ROUTE_RESULT_INDEX } from '../pages/Routes';
-import Result from './Result';
-import useElection from './utils/useElection';
-import './ResultPage.css';
+import { ROUTE_RESULT_INDEX } from './Routes';
+import Result from '../components/Result';
+import useElection from '../components/utils/useElection';
+import './ResultShow.css';
 
-type ResultPageProps = {
+type ResultShowProps = {
   location?: any;
 };
 
-const ResultPage: FC<ResultPageProps> = (props) => {
+const ResultShow: FC<ResultShowProps> = (props) => {
   const { t } = useTranslation();
   //props.location.data = id of the election
   const token = sessionStorage.getItem('token');
@@ -37,8 +37,8 @@ const ResultPage: FC<ResultPageProps> = (props) => {
   );
 };
 
-ResultPage.propTypes = {
+ResultShow.propTypes = {
   location: PropTypes.any,
 };
 
-export default ResultPage;
+export default ResultShow;
